audit: fix regulatory errors, code quality, and expand test coverage
Regulatory correctness:
- Fix PSE risk weights: route to bank table per CRE20.10 (was using sovereign)
- Remove QRRE transactor 0.75 scalar (not in Basel III d424 final)
- Implement short-term bank ECRA risk weights (Table 5, CRE20.17)
- Implement is_rre_secured for defaulted exposures (CRE20.101)
- Add input validation guards for division-by-zero (correlation=1, eir=-1, rho bounds)

Code quality:
- Fix all 12 mypy errors (no-any-return casts, types-PyYAML stub)
- Fix all 70 ruff lint issues (Optional→X|None, str,Enum→StrEnum, unused imports)
- Fix Portfolio.__iter__ conflict with Pydantic BaseModel
- Remove unused jinja2 dependency

Test coverage (42% → 67%, 132 → 206 tests):
- Add tests for IFRS 9 staging, SICR, lifetime PD, scenarios
- Add tests for CECL (PD/LGD, loss-rate, WARM, vintage, DCF, Q-factors)
- Add tests for calibration (binomial, Hosmer-Lemeshow, Spiegelhalter, traffic light, Jeffreys, Brier)
- Add tests for Gaussian copula Monte Carlo simulation
- Strengthen weak ECL assertions from "ecl > 0" to exact expected values
- Add tests for new SA features (short-term ECRA, RRE-secured defaults, PSE routing)

+++ src/creditriskengine/ecl/ifrs9/ttc_to_pit.py
@@ -95,4 +95,4 @@ def estimate_z_factor(
     g_ttc = norm.ppf(max(min(long_run_pd, 0.9999), 0.0001))
     g_obs = norm.ppf(max(min(observed_default_rate, 0.9999), 0.0001))
 
-    return (g_ttc - math.sqrt(1.0 - rho) * g_obs) / math.sqrt(rho)
+    return float((g_ttc - math.sqrt(1.0 - rho) * g_obs) / math.sqrt(rho))
